Добавит модули в файл JS
