+++ js/util.js
@@ -1,0 +1,12 @@
+function getRandomNumber(min, max) {
+  min = Math.ceil(min);
+  max = Math.floor(max);
+  if (min >= max) {
+    throw new Error('Невалидный тип данных');
+  }
+  return Math.floor(Math.random() * (max - min + 1)) + min;
+}
+
+const getRandomArrayElement = (elements) => elements[getRandomNumber(0, elements.length - 1)];
+
+export {getRandomNumber, getRandomArrayElement};
